refactor(network): 更新 RESPParser 中的数据类型定义
- 将 array_size_ 的数据类型从 long long 修改为 int64_t
- 将 bulk_len_ 的数据类型从 long long 修改为 int64_t
- 确保数据类型的统一性和跨平台兼容性

+++ include/network/RESPParser.h
@@ -35,8 +35,8 @@ class RESPParser {
 
  private:
   State state_;
-  long long array_size_;  // 数组总元素数
-  long long bulk_len_;    // 当前正在解析的字符串长度
+  int64_t  array_size_;  // 数组总元素数
+  int64_t bulk_len_;    // 当前正在解析的字符串长度
 };
 
 #endif  // NOVAKV_RESPPARSER_H
